requirements.txt
fastapi==0.115.6
python-multipart==0.0.20
anthropic==0.42.0
openpyxl==3.1.5
supabase==2.11.0
python-dotenv==1.0.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
email-validator==2.3.0
pdfplumber==0.11.4
